comm_quotes_blog/requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.0.7
django-cleanup==5.0.0
django-heroku==0.3.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.7.5
pycodestyle==2.6.0
pylint==2.5.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.1.0
wrapt==1.12.1
